Continued APU.

+++ include/GUI.h
@@ -7,6 +7,7 @@
 #define GUI_HEIGHT 240
 #define GUI_SCALE 2
 
+// #define AUDIO_SAMPLE_RATE 44100.0
 #define AUDIO_SAMPLE_RATE 48000.0
 
 class GUI {
@@ -42,7 +43,7 @@ class GUI {
         SDL_AudioDeviceID audio_device;
         SDL_AudioSpec audio_spec;
 
-        uint16_t volume;
+        uint32_t volume;
 
         // Constructor / Destructor.
         GUI(int width = GUI_WIDTH, int height = GUI_HEIGHT, int scale = GUI_SCALE);
